rename communitieslist to communitylist
