+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import CommunitiesList from "./pages/CommunitiesList";
+import CommunityList from "./pages/CommunityList";
 import CommunitiesDetail from "./pages/CommunitiesDetail";
 import { CreateCommunity } from "./pages/CreateCommunity";
 import { About } from "./pages/About";
@@ -9,7 +9,7 @@ const App: React.FC = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<CommunitiesList />} />
+        <Route path="/" element={<CommunityList />} />
         <Route path="/communities/:id" element={<CommunitiesDetail />} />
         <Route path="/CreateCommunity" element={<CreateCommunity />} />
         <Route path="/about" element={<About />} />
